[refactor]ViewController 에서 vc1,vc2를 각 MainViewController,WeatherDisplayViewController로 변경함
